// autogpt_platform/frontend/src/app/api/__generated__/models/pagination.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */

export interface Pagination {
  /** Total number of items. */
  total_items: number;
  /** Total number of pages. */
  total_pages: number;
  /** Current_page page number. */
  current_page: number;
  /** Number of items per page. */
  page_size: number;
}